separate out dao logic from trading logic
