+++ tests/src/test_trading.py
@@ -1,0 +1,15 @@
+from ...algofin.src.trading import BackTest, Price
+
+
+def test_true():
+    assert True
+
+
+def test_it_runs_everything():
+    BackTest()
+    assert True
+
+
+
+
+
